Revert "rename: rename *& change package path"
This reverts commit 4bfd384acfed6bc421eda2017630c204aeb1187b.
