+++ src/main/java/com/bugflix/weblog/security/CustomUserDetailServiceImpl.java
@@ -1,7 +1,6 @@
-package com.bugflix.weblog.security.service;
+package com.bugflix.weblog.security;
 
-import com.bugflix.weblog.security.domain.CustomUserDetails;
-import com.bugflix.weblog.user.repository.UserRepository;
+import com.bugflix.weblog.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -16,7 +15,7 @@ public class CustomUserDetailServiceImpl implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return new CustomUserDetails(userRepository.findByEmail(username)
+        return new CustomUserDetails(userRepository.findByLoginId(username)
                 .orElseThrow(() -> new UsernameNotFoundException("일치하는 User를 찾을 수 없습니다."))
         );
     }
