+++ src/main/java/com/bugflix/weblog/user/User.java
@@ -1,4 +1,4 @@
-package com.bugflix.weblog.user.domain;
+package com.bugflix.weblog.user;
 
 import com.bugflix.weblog.common.BaseTimeEntity;
 import jakarta.persistence.*;
@@ -11,8 +11,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Table(name = "user_tb")
+@Entity
 @Getter
-@Entity(name = "user_tb")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class User extends BaseTimeEntity {
     @Id
@@ -20,23 +21,19 @@ public class User extends BaseTimeEntity {
     @Column(name = "user_id")
     private Long userId;
 
-    @Column(unique = true)
-    private String email;
+    @Column(unique = true, name = "login_id")
+    private String loginId;
 
     private String password;
 
     private String nickname;
 
-    @JoinColumn(name = "profile_id")
-    @OneToOne
-    private Profile profile;
-
     @OneToMany(mappedBy = "user", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     private List<Authority> roles = new ArrayList<>();
 
     @Builder
-    public User(String email, String password, String nickname) {
-        this.email = email;
+    public User(String loginId, String password, String nickname) {
+        this.loginId = loginId;
         this.password = password;
         this.nickname = nickname;
     }
