+++ src/main/java/com/bugflix/weblog/user/UserController.java
@@ -1,16 +1,11 @@
-package com.bugflix.weblog.user.controller;
+package com.bugflix.weblog.user;
 
-import com.bugflix.weblog.user.service.UserServiceImpl;
-import com.bugflix.weblog.user.dto.SignInRequest;
+import com.bugflix.weblog.security.dto.SignInRequest;
 import com.bugflix.weblog.user.dto.SignUpRequest;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-@Slf4j
-@Tag(name = "User API Endpoint")
 @RestController
 @RequestMapping("/api/users")
 @RequiredArgsConstructor
@@ -20,7 +15,6 @@ public class UserController {
 
     @PostMapping()
     public ResponseEntity<Void> signUp(@RequestBody SignUpRequest signUpRequest) {
-        log.info("회원가입");
         userService.register(signUpRequest);
         return ResponseEntity.ok().build();
     }
@@ -28,11 +22,6 @@ public class UserController {
     @DeleteMapping()
     public ResponseEntity<Void> unregister(@RequestBody SignInRequest signInRequest) throws Exception {
         userService.unregister(signInRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/test")
-    public ResponseEntity<Void> test() {
         return ResponseEntity.ok().build();
     }
